refactor: Header image shown through background-image CSS property
Instead of using an img element in HeaderImage, the component (renamed to BackgroundImage) uses the background-image CSS property as an inline style for the respective div. The image url is passed to BackgroundImage from its parent page component as props.src. Page components import the images directly instead of using image.js as an intermediary.