refactor: Changes a lot to lock Nav on certain pages and hide Menu when clicking new pages
Adds a reducer to nav-slice to help with transparency issues. Each page now sets what the transparency setting should be for that page. The Home and Project pages also  unlock the Nav when rendered now. And Nav the if statements regarding dispatches are more specific so they fire off fewer dispatches and cause less bugs