refactor: makes useScrollPosition agnostic about device size
Scroll and touchmove event listeners are added simulatenously instead of depending on the device size. This fixes the following issues: going from a small to large screen and not having features work because  a scroll listener wasn't added, and using a touch screen laptop but not being able to use touch functionality because a touchmove event listener wasn't added.